refactor(config): reorganize configuration imports and remove obsolete test file
